+++ scripts/update.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 git pull
-systemctl --user stop llama-benchmark.service
-systemctl --user start llama-benchmark.service
+systemctl --user restart llama-benchmark.service
+
+# systemctl --user stop llama-benchmark.service
+# systemctl --user start llama-benchmark.service